Create triangulation.h
